+++ trades-receiver-service/internal/entity/company.go
@@ -2,7 +2,7 @@ package entity
 
 type Company struct {
 	ID     string
-	Cik    int
+	Cik    int64
 	Name   string
 	Ticker string
 }
