+++ trades-receiver-service/internal/entity/insider.go
@@ -2,6 +2,6 @@ package entity
 
 type Insider struct {
 	ID   string
-	Cik  int
+	Cik  int64
 	Name string
 }
